+++ demo/react-ts-webpack/package.json
@@ -15,6 +15,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-loader-spinner": "^6.1.6",
+    "react-router-dom": "^6.26.2",
     "tlsn-js": "../../"
   },
   "devDependencies": {
