+++ demo/react-ts-webpack/src/utils/worker.ts
